requirements.txt
PyQt6>=6.9.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
segyio>=1.9.0
pyyaml>=6.0
openpyxl>=3.0.0
obspy>=1.4.0
datetime>=5.0
tqdm>=4.65.0
scikit-learn>=1.0.0 